refactor: product common으로 이동으로 인해 삭제
