+++ commerce/src/main/java/store/_0982/commerce/CommerceApplication.java
@@ -2,6 +2,7 @@ package store._0982.commerce;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -10,6 +11,10 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableScheduling
 @EnableFeignClients
 @SpringBootApplication
+@EntityScan(basePackages = {
+        "store._0982.commerce.domain",
+        "store._0982.common.domain"
+})
 public class CommerceApplication {
 
     public static void main(String[] args) {
