+++ commerce/src/main/java/store/_0982/commerce/application/grouppurchase/dto/GroupPurchaseDetailInfo.java
@@ -2,7 +2,7 @@ package store._0982.commerce.application.grouppurchase.dto;
 
 import store._0982.commerce.domain.grouppurchase.GroupPurchase;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseStatus;
-import store._0982.commerce.domain.product.ProductCategory;
+import store._0982.common.domain.product.ProductCategory;
 
 import java.time.OffsetDateTime;
 import java.util.UUID;
