+++ commerce/src/main/java/store/_0982/commerce/application/grouppurchase/event/GroupPurchaseParticipatedEvent.java
@@ -1,7 +1,7 @@
 package store._0982.commerce.application.grouppurchase.event;
 
 import store._0982.commerce.domain.grouppurchase.GroupPurchase;
-import store._0982.commerce.domain.product.Product;
+import store._0982.common.domain.product.Product;
 
 public record GroupPurchaseParticipatedEvent (
         GroupPurchase groupPurchase,
