+++ commerce/src/main/java/store/_0982/commerce/application/product/ProductService.java
@@ -9,15 +9,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import store._0982.commerce.application.product.dto.*;
 import store._0982.commerce.application.product.event.ProductUpsertedEvent;
+import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseStatus;
+import store._0982.commerce.domain.product.ProductRepository;
 import store._0982.commerce.domain.product.ProductVectorRepository;
+import store._0982.commerce.exception.CustomErrorCode;
+import store._0982.common.domain.product.Product;
 import store._0982.common.dto.PageResponse;
 import store._0982.common.exception.CustomException;
 import store._0982.common.log.ServiceLog;
-import store._0982.commerce.exception.CustomErrorCode;
-import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
-import store._0982.commerce.domain.product.Product;
-import store._0982.commerce.domain.product.ProductRepository;
 
 import java.util.List;
 import java.util.Optional;
