+++ commerce/src/main/java/store/_0982/commerce/application/product/dto/ProductListInfo.java
@@ -1,7 +1,7 @@
 package store._0982.commerce.application.product.dto;
 
-import store._0982.commerce.domain.product.Product;
-import store._0982.commerce.domain.product.ProductCategory;
+import store._0982.common.domain.product.Product;
+import store._0982.common.domain.product.ProductCategory;
 
 import java.time.OffsetDateTime;
 import java.util.UUID;
