+++ commerce/src/main/java/store/_0982/commerce/application/product/event/ProductUpsertedEvent.java
@@ -1,6 +1,6 @@
 package store._0982.commerce.application.product.event;
 
-import store._0982.commerce.domain.product.Product;
+import store._0982.common.domain.product.Product;
 
 public record ProductUpsertedEvent(
         Product product
