--- commerce/src/main/java/store/_0982/commerce/domain/product/ProductCategory.java
@@ -1,13 +1,0 @@
-package store._0982.commerce.domain.product;
-
-public enum ProductCategory {
-    HOME,        // 생활 & 주방
-    FOOD ,       // 식품 & 간식
-    HEALTH,      // 건강 & 헬스
-    BEAUTY,      // 뷰티
-    FASHION,     // 패션 & 의류
-    ELECTRONICS, // 전자 & 디지털
-    KIDS,        // 유아 & 어린이
-    HOBBY,       // 취미
-    PET          // 반려동물
-}
