+++ commerce/src/main/java/store/_0982/commerce/domain/product/ProductRepository.java
@@ -2,6 +2,7 @@ package store._0982.commerce.domain.product;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import store._0982.common.domain.product.Product;
 
 import java.util.List;
 import java.util.Optional;
