+++ commerce/src/main/java/store/_0982/commerce/infrastructure/product/ProductJpaRepository.java
@@ -2,8 +2,8 @@ package store._0982.commerce.infrastructure.product;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import store._0982.commerce.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
+import store._0982.common.domain.product.Product;
 
 import java.util.List;
 import java.util.Optional;
