+++ commerce/src/test/java/store/_0982/commerce/application/grouppurchase/ParticipateServiceConcurrencyTest.java
@@ -13,8 +13,6 @@ import org.testcontainers.junit.jupiter.Container;
 import store._0982.commerce.domain.grouppurchase.GroupPurchase;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
 import store._0982.commerce.domain.grouppurchase.GroupPurchaseStatus;
-import store._0982.commerce.domain.product.Product;
-import store._0982.commerce.domain.product.ProductCategory;
 import store._0982.commerce.domain.product.ProductRepository;
 import store._0982.commerce.exception.CustomErrorCode;
 import store._0982.commerce.support.BaseConcurrencyTest;
